FIx unit tests, move more read-only functions to RedisCoordinator
