+++ arbnode/seq_coordinator.go
@@ -26,15 +26,6 @@ import (
 	"github.com/offchainlabs/nitro/util/stopwaiter"
 )
 
-const CHOSENSEQ_KEY string = "coordinator.chosen"              // Never overwritten. Expires or released only
-const MSG_COUNT_KEY string = "coordinator.msgCount"            // Only written by sequencer holding CHOSEN key
-const PRIORITIES_KEY string = "coordinator.priorities"         // Read only
-const LIVELINESS_KEY_PREFIX string = "coordinator.liveliness." // Per server. Only written by self
-const MESSAGE_KEY_PREFIX string = "coordinator.msg."           // Per Message. Only written by sequencer holding CHOSEN
-const LIVELINESS_VAL string = "OK"
-const INVALID_VAL string = "INVALID"
-const INVALID_URL string = "<?INVALID-URL?>"
-
 type SeqCoordinator struct {
 	stopwaiter.StopWaiter
 
@@ -170,10 +161,6 @@ func atomicTimeRead(addr *int64) time.Time {
 
 func livelinessKeyFor(url string) string { return LIVELINESS_KEY_PREFIX + url }
 
-func messageKeyFor(pos arbutil.MessageIndex) string {
-	return fmt.Sprintf("%s%d", MESSAGE_KEY_PREFIX, pos)
-}
-
 func execTestPipe(pipe redis.Pipeliner, ctx context.Context) error {
 	cmders, err := pipe.Exec(ctx)
 	if err != nil {
@@ -256,29 +243,6 @@ func (c *SeqCoordinator) chosenOneUpdate(ctx context.Context, msgCountExpected, 
 	}
 	atomicTimeWrite(&c.lockoutUntil, lockoutUntil.Add(-c.config.LockoutSpare))
 	return nil
-}
-
-func (c *SeqCoordinator) getRemoteMsgCountImpl(ctx context.Context, r redis.Cmdable) (arbutil.MessageIndex, error) {
-	resStr, err := r.Get(ctx, MSG_COUNT_KEY).Result()
-	if errors.Is(err, redis.Nil) {
-		return 0, nil
-	}
-	if err != nil {
-		return 0, err
-	}
-	resBytes := []byte(resStr)
-	resBytes, err = c.signer.VerifyMessageSignature(nil, resBytes)
-	if err != nil {
-		return 0, err
-	}
-	if len(resBytes) != 8 {
-		return 0, fmt.Errorf("unexpected msg count value length %v", len(resBytes))
-	}
-	return arbutil.MessageIndex(binary.BigEndian.Uint64(resBytes)), nil
-}
-
-func (c *SeqCoordinator) GetRemoteMsgCount() (arbutil.MessageIndex, error) {
-	return c.getRemoteMsgCountImpl(c.GetContext(), c.client)
 }
 
 func (c *SeqCoordinator) livelinessUpdate(ctx context.Context) error {
