+++ arbnode/seq_coordinator_atomic_test.go
@@ -1,9 +1,6 @@
 // Copyright 2021-2022, Offchain Labs, Inc.
 // For license information, see https://github.com/nitro/blob/master/LICENSE
 
-//go:build redistest
-// +build redistest
-
 package arbnode
 
 import (
@@ -72,7 +69,7 @@ func coordinatorTestThread(ctx context.Context, coord *SeqCoordinator, data *Coo
 			timeLaunching := time.Now()
 			// didn't sequence.. should we have succeeded?
 			if timeLaunching.Before(holdingLockout) {
-				execError = fmt.Errorf("failed while holding lock %s err %w", coord.config.MyUrl, err)
+				execError = fmt.Errorf("failed while holding lock %s err %w", coord.config.MyUrl(), err)
 				break
 			}
 		}
@@ -82,9 +79,9 @@ func coordinatorTestThread(ctx context.Context, coord *SeqCoordinator, data *Coo
 				continue
 			}
 			if data.sequencer[i] != "" {
-				execError = fmt.Errorf("two sequencers for same msg: submsg %d, success for %s, %s", i, data.sequencer[i], coord.config.MyUrl)
+				execError = fmt.Errorf("two sequencers for same msg: submsg %d, success for %s, %s", i, data.sequencer[i], coord.config.MyUrl())
 			}
-			data.sequencer[i] = coord.config.MyUrl
+			data.sequencer[i] = coord.config.MyUrl()
 		}
 		if execError != nil {
 			data.err = execError
@@ -113,14 +110,19 @@ func TestRedisSeqCoordinatorAtomic(t *testing.T) {
 
 	redisClient, err := redisutil.RedisClientFromURL(redisutil.GetTestRedisURL(t))
 	Require(t, err)
+	if redisClient == nil {
+		t.Fatal("redisClient is nil")
+	}
 
 	for i := 0; i < NumOfThreads; i++ {
 		config := coordConfig
-		config.MyUrl = fmt.Sprint(i)
+		config.MyUrlImpl = fmt.Sprint(i)
+		redisCoordinator, err := NewRedisCoordinator(config.RedisUrl, config.MyUrl())
+		Require(t, err)
 		coordinator := &SeqCoordinator{
-			client: redisClient,
-			config: config,
-			signer: nullSigner,
+			RedisCoordinator: *redisCoordinator,
+			config:           config,
+			signer:           nullSigner,
 		}
 		go coordinatorTestThread(ctx, coordinator, &testData)
 	}
